Change: moved recording tests to their own file
